+++ src/classes/CustomLibMetadata/SyntaxAnalyzer.ts
@@ -171,11 +171,13 @@ export class SyntaxAnalyzer {
 		return completionItems;
 	}
 
-	public static gerCurrentClass() {
+	public static gerCurrentClass(documentText?: string) {
 		let returnClassName;
-		if (vscode.window.activeTextEditor) {
-			const documentText = vscode.window.activeTextEditor.document.getText();
-			const rCurrentClass = /(?<=.*\..*\(\").*(?=\")/;
+		if (!documentText && vscode.window.activeTextEditor) {
+			documentText = vscode.window.activeTextEditor.document.getText();
+		}
+		if (documentText) {
+			const rCurrentClass = /(?<=.*\..*(extend|declareStaticClass)\(\").*(?=\")/;
 			const rCurrentClassResults = rCurrentClass.exec(documentText);
 			if (rCurrentClassResults) {
 				returnClassName = rCurrentClassResults[0];