+++ CHANGELOG.md
@@ -9,3 +9,10 @@
 * Patch with bugfixes
 * Add support of Completion item generation depending on this.byId (previously worked with this.getView().byId only)
 * Now the extension is initialized only if there is manifest.json found in the workspace
+
+## 0.7.0 (05-12-2019)
+* "Insert Custom Class name" command added
+* Now code templates are automatically added to new .js files
+* Now on file renaming all occurances of previous class name are replaced with new class name (for now works for .js files only)
+* Bugfixes for method definition finder
+* Partial support for ES6 added